pyproject.toml
[project]
name = "obs-agents-core"
version = "0.1.0"
description = "Autonomous agents for Obsidian knowledge contents."
authors = [{ name = "akitorahayashi", email = "atora.has75@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django>=5.2.5,<6.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "gunicorn>=22.0.0,<23.0.0",
    "dj-database-url>=2.2.0,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "djangorestframework>=3.15.0,<4.0.0",
    "django-cors-headers>=4.3.0,<5.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.9",
    "ipython>=9.4.0",
    "pytest-django>=4.11.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.7",
    "testcontainers[postgres]>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[tool.black]
target-version = ['py312']
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.venv
    | venv
    | env
    | build
    | dist
    | .*/(bin|lib|share|etc)
    | .*/pyvenv\.cfg
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
exclude = [
	"__pycache__",
	"migrations",
	"db.sqlite3",
	"*.pyc",
	".venv",
	"venv",
	"env",
	"*/bin/**",
	"*/lib/**",
	"*/share/**",
	"*/etc/**",
	"*/pyvenv.cfg"
]

[tool.ruff.lint]
select = ["E", "F", "I"]
# E501 (line too long) is ignored
ignore = ["E501"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
